requirements.txt
BiblioPixel<3.0
BiblioPixelAnimations==3.20170531.153148
Pillow
pyserial
